server/requirements.txt
Flask
flask_cors
flask-sqlalchemy
flask_sqlalchemy
pymysql
werkzeug
gunicorn
